third_party/blink/web_tests/platform/generic/fast/js/lastModified-expected.txt

This test verifies that the lastModified property exists on the document object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'lastModified' in document is true
PASS successfullyParsed is true

TEST COMPLETE
